fix(service): removed package-lock and organized and modified openrouter service #88
